Conflicts:
	Makefile

+++ common.h
@@ -1,8 +1,10 @@
 #ifndef _COMMON_H_
 #define _COMMON_H_
 
-#define P(args...) printf("[%d] ", ID); printf(args); printf("\n"); fflush(stdout)
+#include <inttypes.h>
+
+#define P(args...) printf("[%02d] ", ID); printf(args); printf("\n"); fflush(stdout)
 #define PRINT P
 
-extern int ID;
+extern uint8_t ID;
 #endif
